Fix build error (printf argument mismatch)
Resolve these build errors:
gsttividenc1.c:1375: error: format '%lu' expects type 'long unsigned int', but argument 2 has type 'guint'
gsttividenc1.c:1375: error: format '%lu' expects type 'long unsigned int', but argument 3 has type 'unsigned int'

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@880 24075187-2e39-4e88-bbb8-bc8aa768f540

+++ gstreamer_ti/ti_build/ticodecplugin/src/gsttividenc1.c
@@ -1374,7 +1374,8 @@ gst_tividenc1_encode(GstTIVidenc1 *videnc1, GstBuffer *inBuf,
     if (GST_BUFFER_SIZE(inBuf) != videnc1->upstreamBufSize) {
         GST_ELEMENT_ERROR(videnc1, RESOURCE, NO_SPACE_LEFT,
         ("input buffer is an invalid size (%lu != %lu)\n",
-        (guint32)GST_BUFFER_SIZE(inBuf), (guint32)videnc1->upstreamBufSize),
+        (unsigned long)GST_BUFFER_SIZE(inBuf),
+        (unsigned long)videnc1->upstreamBufSize),
         (NULL));
         goto exit_fail;
     }
